SpeedModifierCommand addition
19:30 CST

New command: SpeedModifierCommand
New button: SpeedModifierButton

These two additions allow the switching between two speedModifier values
in the SwerveModule.

Passing of Info:
RobotContainer: checks if 0 XB_LB pressed.
If yes: sends 0.5 to SpeedModifierCommand
If no: sends 0.8 to SpeedModifierCommand
In the SpeedModifierCommand, it sets the DriveSubsystem's speedModifier
to it's speedModifier value.